+++ training_log_app/src/ui/create_new_user_view.py
@@ -1,5 +1,5 @@
 from tkinter import ttk, constants
-from services.log_entry_service import log_entry_service, UsernameAlreadyInUseError
+from services.log_entry_service import log_entry_service, UsernameAlreadyInUseError, InvalidInputError
 
 
 class CreateNewUserView:
@@ -9,6 +9,7 @@ class CreateNewUserView:
         self._root = root
         self._show_login_view = login_view
         self._frame = None
+        self.error_label = None
 
         self._initialize()
 
@@ -22,39 +23,62 @@ class CreateNewUserView:
         username = self._username_entry.get()
         password = self._password_entry.get()
 
+
         try:
             log_entry_service.create_new_user(username, password)
-            print('user created successfully')
             self._show_login_view()
+
         except UsernameAlreadyInUseError:
-            print('Username already taken')
+            if self.error_label:
+                self.error_label.grid_remove()
+            self.error_label = ttk.Label(master=self._create_new_user_frame,
+                                        text='Username already taken!',
+                                        foreground='red')
+            self.error_label.grid(row=3, column=0, columnspan=2)
+
+        except InvalidInputError:
+            if self.error_label:
+                self.error_label.grid_remove()
+            self.error_label = ttk.Label(master=self._create_new_user_frame,
+                                        text='Invalid username or password!',
+                                        foreground='red')
+            self.error_label.grid(row=3, column=0, columnspan=2)
+
+    def _handle_go_back(self):
+        self._show_login_view()
 
     def _initialize(self):
-        # initialize window
-
         self._frame = ttk.Frame(master=self._root)
         heading_label = ttk.Label(master=self._frame, text='Create new user')
 
-        username_label = ttk.Label(master=self._frame, text='Username:')
-        self._username_entry = ttk.Entry(master=self._frame)
+        self._create_new_user_frame = ttk.Labelframe(master=self._frame,
+                                                     text='Create a new user',
+                                                     labelanchor='n')
 
-        password_label = ttk.Label(master=self._frame, text='Password:')
-        self._password_entry = ttk.Entry(master=self._frame)
+
+        username_label = ttk.Label(master=self._create_new_user_frame, text='Username:')
+        self._username_entry = ttk.Entry(master=self._create_new_user_frame)
+
+        password_label = ttk.Label(master=self._create_new_user_frame, text='Password:')
+        self._password_entry = ttk.Entry(master=self._create_new_user_frame)
 
         create_button = ttk.Button(master=self._frame,
                                    text='Create user',
                                    command=self._handle_create_new_user)
-
-        # make go back button
+        
+        back_button = ttk.Button(master=self._frame,
+                                 text='Back',
+                                 command=self._handle_go_back)
 
         # build the ui
+        self._create_new_user_frame.grid(columnspan=2, padx=5, pady=5)
 
-        heading_label.grid(row=0, column=0)
+        username_label.grid(row=1, column=0, padx=5, pady=5)
+        self._username_entry.grid(row=1, column=1, padx=5, pady=5)
 
-        username_label.grid(row=1, column=0)
-        self._username_entry.grid(row=1, column=1)
+        password_label.grid(row=2, column=0, padx=5, pady=5)
+        self._password_entry.grid(row=2, column=1, padx=5, pady=5)
 
-        password_label.grid(row=2, column=0)
-        self._password_entry.grid(row=2, column=1)
+        create_button.grid(row=3, column=1, padx=5, pady=5)
 
-        create_button.grid(row=3, column=1)
+        back_button.grid(row=3, column=0, padx=5, pady=5)